+++ src/ItunesMoviePriceTracker.Repository/Repositories/IMoviePriceRepository.cs
@@ -4,8 +4,8 @@ namespace ItunesMoviePriceTracker.Repository.Repositories;
 
 public interface IMoviePriceRepository
 {
-    Task<IEnumerable<MoviePrice>> GetByMovieIdAsync(int trackId);
-    Task<MoviePrice?> GetLatestByMovieIdAsync(int trackId);
+    Task<IEnumerable<MoviePrice>> GetByMovieIdAsync(long trackId);
+    Task<MoviePrice?> GetLatestByMovieIdAsync(long trackId);
     Task<IEnumerable<MoviePrice>> GetTodaysChangesAsync();
     Task AddAsync(MoviePrice moviePrice);
 }
