+++ src/ItunesMoviePriceTracker.Services/Implementation/MovieService.cs
@@ -15,19 +15,19 @@ public class MovieService(IMovieRepository movieRepository,
         return movies.Select(MovieMapper.ToDto);
     }
 
-    public async Task<MovieDto?> GetMovieAsync(int trackId)
+    public async Task<MovieDto?> GetMovieAsync(long trackId)
     {
         var movie = await movieRepository.GetByIdAsync(trackId);
         return movie is null ? null : MovieMapper.ToDto(movie);
     }
 
-    public async Task<MovieDto?> GetMovieWithPricesAsync(int trackId)
+    public async Task<MovieDto?> GetMovieWithPricesAsync(long trackId)
     {
         var movie = await movieRepository.GetByIdWithPricesAsync(trackId);
         return movie is null ? null : MovieMapper.ToDto(movie);
     }
 
-    public async Task<bool> AddMovieAsync(int trackId, decimal? watchPrice)
+    public async Task<bool> AddMovieAsync(long trackId, decimal? watchPrice)
     {
         if (await movieRepository.ExistsAsync(trackId)) return false;
 
@@ -44,7 +44,7 @@ public class MovieService(IMovieRepository movieRepository,
         return true;
     }
 
-    public async Task UpdateWatchPriceAsync(int trackId, decimal? watchPrice)
+    public async Task UpdateWatchPriceAsync(long trackId, decimal? watchPrice)
     {
         var movie = await movieRepository.GetByIdAsync(trackId);
         if (movie is null) return;
@@ -53,7 +53,7 @@ public class MovieService(IMovieRepository movieRepository,
         await movieRepository.UpdateAsync(movie);
     }
 
-    public async Task<bool> DeleteMovieAsync(int trackId)
+    public async Task<bool> DeleteMovieAsync(long trackId)
     {
         if (!await movieRepository.ExistsAsync(trackId)) return false;
         await movieRepository.DeleteAsync(trackId);
