+++ src/ItunesMoviePriceTracker.Services/Interfaces/IItunesApiService.cs
@@ -4,5 +4,5 @@ namespace ItunesMoviePriceTracker.Services.Interfaces;
 
 public interface IItunesApiService
 {
-    Task<ItunesMovieResult?> FetchMovieAsync(int trackId);
+    Task<ItunesMovieResult?> FetchMovieAsync(long trackId);
 }
