+++ src/ItunesMoviePriceTracker.Services/Interfaces/IMovieService.cs
@@ -5,9 +5,9 @@ namespace ItunesMoviePriceTracker.Services.Interfaces;
 public interface IMovieService
 {
     Task<IEnumerable<MovieDto>> GetAllMoviesAsync();
-    Task<MovieDto?> GetMovieAsync(int trackId);
-    Task<MovieDto?> GetMovieWithPricesAsync(int trackId);
-    Task<bool> AddMovieAsync(int trackId, decimal? watchPrice);
-    Task UpdateWatchPriceAsync(int trackId, decimal? watchPrice);
-    Task<bool> DeleteMovieAsync(int trackId);
+    Task<MovieDto?> GetMovieAsync(long trackId);
+    Task<MovieDto?> GetMovieWithPricesAsync(long trackId);
+    Task<bool> AddMovieAsync(long trackId, decimal? watchPrice);
+    Task UpdateWatchPriceAsync(long trackId, decimal? watchPrice);
+    Task<bool> DeleteMovieAsync(long trackId);
 }
