+++ src/ItunesMoviePriceTracker.Shared/DTOs/MoviePriceDto.cs
@@ -5,6 +5,6 @@ public class MoviePriceDto
     public int Id { get; set; }
     public DateTime Date { get; set; }
     public decimal Price { get; set; }
-    public int MovieTrackId { get; set; }
+    public long MovieTrackId { get; set; }
     public string? TrackName { get; set; }
 }
